+++ packages/console/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @hot-updater/console
 
+## 0.31.4
+
+### Patch Changes
+
+- @hot-updater/bsdiff@0.31.4
+- @hot-updater/server@0.31.4
+
+## 0.31.3
+
+### Patch Changes
+
+- @hot-updater/bsdiff@0.31.3
+- @hot-updater/server@0.31.3
+
 ## 0.31.2
 
 ### Patch Changes