+++ plugins/expo/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @hot-updater/expo
 
+## 0.31.4
+
+### Patch Changes
+
+- @hot-updater/cli-tools@0.31.4
+- @hot-updater/bare@0.31.4
+- @hot-updater/plugin-core@0.31.4
+
+## 0.31.3
+
+### Patch Changes
+
+- @hot-updater/cli-tools@0.31.3
+- @hot-updater/bare@0.31.3
+- @hot-updater/plugin-core@0.31.3
+
 ## 0.31.2
 
 ### Patch Changes